requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==23.0.0
mercado-pago==2.3.2
python-decouple==3.8